132486 use tagEditHandler in pageControls
